requirements.txt
sqlalchemy
s3fs
python-dotenv
pandas